move all backend files and folders to /backend and create client dir
