+++ src/backend/api/index.ts
@@ -8,7 +8,7 @@ import BodyParser from "body-parser";
 import GitHubAPI from "../../config/github-api";
 import { Installation, Subscription } from "../models";
 import verifyInstallation from "../jira/verify-installation";
-import logMiddleware from "../../middleware/log-middleware";
+import logMiddleware from "../middleware/log-middleware";
 import JiraClient from "../models/jira-client";
 import uninstall from "../jira/uninstall";
 import { serializeJiraInstallation, serializeSubscription } from "./serializers";
