+++ src/backend/app.ts
@@ -5,29 +5,29 @@ import cookieSession from "cookie-session";
 import csrf from "csurf";
 import * as Sentry from "@sentry/node";
 import hbs from "hbs";
-import GithubOAuth from "./github-oauth";
-import getGitHubSetup from "./get-github-setup";
-import postGitHubSetup from "./post-github-setup";
-import getGitHubConfiguration from "./get-github-configuration";
-import postGitHubConfiguration from "./post-github-configuration";
-import listGitHubInstallations from "./list-github-installations";
-import getGitHubSubscriptions from "./get-github-subscriptions";
-import deleteGitHubSubscription from "./delete-github-subscription";
-import getJiraConfiguration from "./get-jira-configuration";
-import deleteJiraConfiguration from "./delete-jira-configuration";
-import getGithubClientMiddleware from "./github-client-middleware";
-import verifyJiraMiddleware from "./verify-jira-middleware";
-import getJiraConnect from "../backend/jira/connect";
-import postJiraDisable from "../backend/jira/disable";
-import postJiraEnable from "../backend/jira/enable";
-import postJiraInstall from "../backend/jira/install";
-import postJiraUninstall from "../backend/jira/uninstall";
-import jiraAuthenticate from "../backend/jira/authenticate";
-import retrySync from "./retry-sync";
-import getMaintenance from "./get-maintenance";
-import api from "../backend/api";
-import healthcheck from "./healthcheck";
-import logMiddleware from "../middleware/log-middleware";
+import GithubOAuth from "./middleware/github-oauth";
+import getGitHubSetup from "./middleware/get-github-setup";
+import postGitHubSetup from "./middleware/post-github-setup";
+import getGitHubConfiguration from "./middleware/get-github-configuration";
+import postGitHubConfiguration from "./middleware/post-github-configuration";
+import listGitHubInstallations from "./middleware/list-github-installations";
+import getGitHubSubscriptions from "./middleware/get-github-subscriptions";
+import deleteGitHubSubscription from "./middleware/delete-github-subscription";
+import getJiraConfiguration from "./middleware/get-jira-configuration";
+import deleteJiraConfiguration from "./middleware/delete-jira-configuration";
+import getGithubClientMiddleware from "./middleware/github-client-middleware";
+import verifyJiraMiddleware from "./middleware/verify-jira-middleware";
+import getJiraConnect from "./jira/connect";
+import postJiraDisable from "./jira/disable";
+import postJiraEnable from "./jira/enable";
+import postJiraInstall from "./jira/install";
+import postJiraUninstall from "./jira/uninstall";
+import jiraAuthenticate from "./jira/authenticate";
+import retrySync from "./middleware/retry-sync";
+import getMaintenance from "./middleware/get-maintenance";
+import api from "./api";
+import healthcheck from "./middleware/healthcheck";
+import logMiddleware from "./middleware/log-middleware";
 import { App } from "@octokit/app";
 import statsd, { elapsedTimeMetrics } from "../config/statsd";
 import { metricError } from "../config/metric-names";
