+++ src/backend/middleware/delete-github-subscription.ts
@@ -1,6 +1,6 @@
-import { ActionFromSubscription, ActionSource, ActionType } from "../backend/proto/v0/action";
-import { submitProto } from "../backend/tracking";
-import { Installation, Subscription } from "../backend/models";
+import { ActionFromSubscription, ActionSource, ActionType } from "../proto/v0/action";
+import { submitProto } from "../tracking";
+import { Installation, Subscription } from "../models";
 import { Request, Response } from "express";
 
 /**
