+++ src/backend/middleware/delete-jira-configuration.ts
@@ -1,7 +1,7 @@
-import { ActionFromSubscription, ActionSource, ActionType } from "../backend/proto/v0/action";
-import { submitProto } from "../backend/tracking";
-import { Subscription } from "../backend/models";
-import getJiraClient from "../backend/jira/client";
+import { ActionFromSubscription, ActionSource, ActionType } from "../proto/v0/action";
+import { submitProto } from "../tracking";
+import { Subscription } from "../models";
+import getJiraClient from "../jira/client";
 import { Request, Response } from "express";
 
 /**
