+++ src/backend/middleware/get-github-configuration.ts
@@ -1,9 +1,9 @@
 import JWT from "atlassian-jwt";
-import { Installation, Subscription } from "../backend/models";
+import { Installation, Subscription } from "../models";
 import { NextFunction, Request, Response } from "express";
-import { getJiraMarketplaceUrl } from "../common/getUrl";
-import enhanceOctokit from "../config/enhance-octokit";
-import app from "../backend/worker/app";
+import { getJiraMarketplaceUrl } from "../../common/getUrl";
+import enhanceOctokit from "../../config/enhance-octokit";
+import app from "../worker/app";
 import { getInstallation } from "./get-jira-configuration";
 
 const getConnectedStatus = (
