+++ src/backend/middleware/retry-sync.ts
@@ -1,4 +1,4 @@
-import { Subscription } from "../backend/models";
+import { Subscription } from "../models";
 import * as Sentry from "@sentry/node";
 import { NextFunction, Request, Response } from "express";
