--- src/interfaces/github.ts
@@ -1,9 +1,0 @@
-export interface GitHubPullRequest {
-	head: {
-		sha: string;
-		repo: {
-			url: string;
-		};
-		ref: string;
-	};
-}
