+++ test/unit/frontend/delete-jira-configuration.test.ts
@@ -3,7 +3,7 @@ import testTracking from "../../setup/tracking";
 import nock from "nock";
 import { Installation, Subscription } from "../../../src/backend/models";
 import { mocked } from "ts-jest/utils";
-import deleteConfig from "../../../src/frontend/delete-jira-configuration";
+import deleteConfig from "../../../src/backend/middleware/delete-jira-configuration";
 
 jest.mock("../../../src/backend/models");
